requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
execnet==2.1.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-rerunfailures==15.1
pytest-xdist==3.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
